+++ Changelog.md
@@ -40,12 +40,15 @@ struct and can be applied at once. When `serde-types` feature is enabled, config
 - [#924]: (breaking change) Split `SyntaxError::UnclosedPIOrXmlDecl` into `UnclosedPI` and
   `UnclosedXmlDecl` for more precise error reporting.
 - [#924]: (breaking change) `Parser::eof_error` now takes `&self` and content `&[u8]` parameters.
+- [#926]: (breaking change) Split `SyntaxError::UnclosedTag` into `UnclosedTag`,
+  `UnclosedSingleQuotedAttributeValue` and `UnclosedDoubleQuotedAttributeValue` for more precise error reporting.
 
 [#846]: https://github.com/tafia/quick-xml/issues/846
 [#908]: https://github.com/tafia/quick-xml/pull/908
 [#913]: https://github.com/tafia/quick-xml/pull/913
 [#923]: https://github.com/tafia/quick-xml/issues/923
 [#924]: https://github.com/tafia/quick-xml/pull/924
+[#926]: https://github.com/tafia/quick-xml/issues/926
 [#929]: https://github.com/tafia/quick-xml/pull/929